// src/app/models/homeView.ts
export interface HomeView {
  welcomeEN: string
  welcomeES: string
  welcomeDE: string
  welcomeFR: string
  welcomeIT: string
  alcoholicBtnEN: string
  alcoholicBtnES: string
  alcoholicBtnDE: string
  alcoholicBtnFR: string
  alcoholicBtnIT: string
  nonAlcoholicBtnEN: string
  nonAlcoholicBtnES: string
  nonAlcoholicBtnDE: string
  nonAlcoholicBtnFR: string
  nonAlcoholicBtnIT: string
  randomBtnEN: string
  randomBtnES: string
  randomBtnDE: string
  randomBtnFR: string
  randomBtnIT: string
}